Added controls to edit page (+ other changes)
Added controls to edit page (they only change how blue the color is
currently)

Updated UI theme to be the grey buttons on older android phones and the
newer "Holo" theme on newer Android phones. This involved switching the
target API back to Android-18 and creating separate styles.xml files for
the different API levels.

Added a copy-constructor to ColorRecord
Added setInt to ColorRecord to set the color based on an int.